--- src/fab/index.ts
@@ -1,6 +1,0 @@
-import { withTheme } from '../config';
-import FAB, { FABProps } from './FAB';
-
-export type { FABProps };
-
-export default withTheme(FAB, 'floatingActionButton');
