+++ src/index.ts
@@ -9,8 +9,6 @@ import Icon, { IconProps } from './icons/Icon';
 import ListItem, { ListItemProps } from './list/ListItem';
 import SocialIcon, { SocialIconProps } from './social/SocialIcon';
 import Overlay, { OverlayProps } from './overlay/Overlay';
-import LinearProgress, { LinearProgressProps } from './linearProgress';
-import FAB, { FABProps } from './fab';
 
 // Utilities
 import SearchBar, { SearchBarProps } from './searchbar/SearchBar';
@@ -24,6 +22,7 @@ import Divider, { DividerProps } from './divider/Divider';
 import Slider, { SliderProps } from './slider/Slider';
 import ButtonGroup, { ButtonGroupProps } from './buttons/ButtonGroup';
 import Image, { ImageProps } from './image/Image';
+import FAB, { FABProps } from './buttons/FAB';
 // Productivity
 import Card, { CardProps } from './card/Card';
 import Tile, { TileProps } from './tile/Tile';
@@ -33,6 +32,7 @@ import Header, { HeaderProps } from './header/Header';
 import PricingCard, { PricingCardProps } from './pricing/PricingCard';
 import Tooltip, { TooltipProps } from './tooltip/Tooltip';
 import BottomSheet, { BottomSheetProps } from './bottomSheet/BottomSheet';
+import LinearProgress, { LinearProgressProps } from './linearProgress';
 import {
   AirbnbRating as BaseAirbnbRating,
   Rating as BaseRating,
